# Feat: optional indexing and caching of `site.pages`