--- a/docs/content/configuration.md
+++ b/docs/content/configuration.md
@@ -17,6 +17,7 @@ title = "Bartholomew"
 base_url = "http://localhost:3000"
 about = "This site is generated with Bartholomew, the Spin micro-CMS. And this message is in site.toml."
 theme = "fermyon"
+index_site_pages = ["main"]
 enable_shortcodes = false
 
 [extra]
@@ -34,6 +35,7 @@ It has a few pre-defined fields:
 - base_url: a base URL that templates can use to construct full URLs to content. This can be overridden by setting the `-e BASE_URL="https://example.com"` environment variable for Spin.
 - about: a brief description of the site
 - theme: the name of the theme for the website from the `/themes/` folder
+- index_site_pages: A list of templates that require `site.pages` to be populated.
 - enable_shortcodes: Allows addition of shortcodes in the markdown content using rhai scripts. Defaults to false.
 
 You can define your own fields in the `[extra]` section. Anything in `[extra]` is not