This is a ***breaking*** change with a change required in `site.toml` to fix.

Initial implementation of having the ability to index content for `site.pages` for only required templates. Implement the cache proposed in #15 for `site.pages` to improve performance.

Optional indexing improves the performance of pages that do not depend on `site.toml` while caching improves the performance of any that uses it.

The templates for which `site.pages` list is required can be specified in `site.toml` as

```
.
.
index_site_pages  = ["main"]

[extra] 
.
.
```

TODO: Docs. 

Some initial performance results for tests on fermyon.com

## Optional indexing

**This branch**

```
bombardier -c 1 -n 1000 http://localhost:3000/blog/underprovisioning-cloud-services
Bombarding http://localhost:3000/blog/underprovisioning-cloud-services with 1000 request(s) using 1 connection(s)
 1000 / 1000 [=========================================================] 100.00% 76/s 13s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec        77.46      26.13     261.02
  Latency       12.96ms   829.78us    23.75ms
  HTTP codes:
    1xx - 0, 2xx - 1000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     3.43MB/s
```

**main branch**

```
bombardier -c 1 -n 1000 http://localhost:3000/blog/underprovisioning-cloud-services
Bombarding http://localhost:3000/blog/underprovisioning-cloud-services with 1000 request(s) using 1 connection(s)
 1000 / 1000 [=========================================================] 100.00% 45/s 21s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec        45.79       4.47      69.55
  Latency       21.85ms     0.96ms    35.97ms
  HTTP codes:
    1xx - 0, 2xx - 1000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     2.04MB/s
```

## Caching `site.pages`
**This branch**
```
bombardier -c 1 -n 1000 http://localhost:3000/blog/index
Bombarding http://localhost:3000/blog/index with 1000 request(s) using 1 connection(s)
 1000 / 1000 [=========================================================] 100.00% 51/s 19s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec        52.47      21.15     620.55
  Latency       19.25ms     1.77ms    64.88ms
  HTTP codes:
    1xx - 0, 2xx - 1000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     5.07MB/s
```
**main branch**
```
bombardier -c 1 -n 1000 http://localhost:3000/blog/index
Bombarding http://localhost:3000/blog/index with 1000 request(s) using 1 connection(s)
 1000 / 1000 [=========================================================] 100.00% 32/s 30s
Done!
Statistics        Avg      Stdev        Max
  Reqs/sec        33.00      13.88      65.60
  Latency       30.29ms     4.16ms    81.09ms
  HTTP codes:
    1xx - 0, 2xx - 1000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     3.22MB/s
```

Note: To test locally on fermyon.com add the following to `site.toml` when using `bartholomew.wasm` module from this branch to test caching with either homepage or blog page.

```
index_site_pages = ["home", "main"]
```